+++ Adapter/GUI/gui-simulator/Drivers/simulator.hpp
@@ -22,11 +22,11 @@ void simulator_event_Handler();
 uint8_t simulator_is_running();
 
     /********************LCD驱动接口***********************/
-void lcd_write_data(uint16_t page, uint16_t column, uint8_t data);
+void lcd_write_data(uint16_t page, uint16_t column, uint16_t data);
 
-    void lcd_write_page(uint8_t page, const uint8_t*buf);
+    void lcd_write_page(uint16_t page, const uint16_t*buf);
 
-void lcd_full_flush(const uint8_t *buf);// 刷新整个屏幕
+void lcd_full_flush(const uint16_t *buf);// 刷新整个屏幕
 
 void LCD_Clear();
